--- src/utils/decode.ts
@@ -1,6 +1,0 @@
-export const decodeText = (text: string): string => {
-  return (
-    new DOMParser().parseFromString(decodeURI(text), "text/html")
-      .documentElement.textContent || ""
-  );
-};
